+++ src/commonTest/kotlin/de/tsenger/vdstools/dissect/SealDissectorCommonTest.kt
@@ -14,8 +14,11 @@ class SealDissectorCommonTest {
         <?xml version="1.0" encoding="UTF-8"?>
         <profile>
             <profileNumber>AABBCCDD11223344AABBCCDD11223344</profileNumber>
+            <versionTR>0.9</versionTR>
             <profileName>TEST_V9_PROFILE</profileName>
             <creator>Test</creator>
+            <validFromPresent>false</validFromPresent>
+            <validToPresent>false</validToPresent>
             <entry tag="10">
                 <name>SURNAME</name>
                 <description>Familienname</description>