+++ src/commonTest/kotlin/de/tsenger/vdstools/vds/VdsAdministrativeDocumentsV9CommonTest.kt
@@ -28,8 +28,11 @@ class VdsAdministrativeDocumentsV9CommonTest {
         <?xml version="1.0" encoding="UTF-8"?>
         <profile>
             <profileNumber>$testUuid</profileNumber>
+            <versionTR>0.9</versionTR>
             <profileName>TEST_V9_PROFILE</profileName>
             <creator>Test</creator>
+            <validFromPresent>false</validFromPresent>
+            <validToPresent>false</validToPresent>
             <entry tag="10">
                 <name>SURNAME</name>
                 <description>Familienname</description>